LayoutTests/http/tests/site-isolation/pagehide-event-expected.txt

Tests that pagehide is fired when site isolation is enabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.data is "pagehide"
PASS successfullyParsed is true

TEST COMPLETE
